// src/app/shared/interfaces/details/details.ts
export interface Details {
    message: string;
    product: Product;
}

export interface Product {
    rateAvg:            number;
    rateCount:          number;
    _id:                string;
    title:              string;
    slug:               string;
    description:        string;
    imgCover:           string;
    images:             string[];
    price:              number;
    priceAfterDiscount: number;
    quantity:           number;
    category:           string;
    occasion:           string;
    createdAt:          Date;
    updatedAt:          Date;
    __v:                number;
    isSuperAdmin:       boolean;
    sold:               number;
    reviews:            any[];
    id:                 string;
}